Configure arc UserConfig by creating temp ini file instead of using setter methods
It seems the ARC library is buggy when using a proxy file in a location
that is not the default. There is a setter method `ProxyPath(...)`, but
when using this it still seems to try and use the default location to
find the proxy file.

The default location is `x509up_u` with the users's UID appended to it; so
hardcoding the proxy file to this path is not an option. Working out the
user's UID and doing it that way seems to fragile and suspect to changes
in the ARC client.