Preserve a single file/directory on job completion
Since it takes so long for the ARC job status to change to completed,
I had not noticed that all files except std{out,err}.txt are deleted
on job completion.

A new config option is added to specify the name of the file/directory
to download in the call to `save_job_outputs`.

Also added a new example for downloading job outputs